Bug fix in GitRepo: output directory may not exist if compilation fails

+++ GitRepoTracker/Git/GitRepo.cs
@@ -154,17 +154,20 @@ namespace GitRepoTracker
                 outputFolder = testOutputFolder + "\\bin\\Debug\\net6";
 
             //Clean files
-            List<string> allowedExtensions = new List<string> { ".json", ".dll", ".pdb", ".exe" };
-            foreach (string file in Directory.GetFiles(outputFolder))
+            if (Directory.Exists(outputFolder))
             {
-                string fileExtension = Path.GetExtension(file);
-                if (!allowedExtensions.Contains(fileExtension))
-                    File.Delete(file);
-            }
-            //Clean directories
-            foreach (string dir in Directory.GetDirectories(outputFolder))
-            {
-                Directory.Delete(dir, true);
+                List<string> allowedExtensions = new List<string> { ".json", ".dll", ".pdb", ".exe" };
+                foreach (string file in Directory.GetFiles(outputFolder))
+                {
+                    string fileExtension = Path.GetExtension(file);
+                    if (!allowedExtensions.Contains(fileExtension))
+                        File.Delete(file);
+                }
+                //Clean directories
+                foreach (string dir in Directory.GetDirectories(outputFolder))
+                {
+                    Directory.Delete(dir, true);
+                }
             }
 
             string output = RunCommand(args, "dotnet", true);
